+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## Unreleased
+## Capybara Test Helpers 1.0.4 (2020-11-26) ##
 
-* Allow to use aliases in `assert_` methods as well, for consistency.
+* Add `have_no` alias for `have_no_selector`.
+* Add `has_no?` alias for `has_no_selector?`.
 * Remove internal method `wrap_test_helper`.
+* Allow to use locator aliases in `assert_` methods as well, for consistency.
 
 ## Capybara Test Helpers 1.0.3 (2020-11-24) ##